implemented rename and duplicate functionalities and fixed locked for modification bug

+++ UiDesktopApp2/DataAccess/Entities/Result.cs
@@ -12,7 +12,7 @@ namespace UiDesktopApp2.DataAccess.Entities
         [Key]
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public int AvailableTimePerImageVariant { get; set; }
+        public double AvailableTimePerImageVariant { get; set; }
         public int SubjectId { get; set; }
         public Subject Subject { get; set; }
         public int TestId { get; set; }
